refactor: split handlers
